gui: improve set_key and info
we dont use a info messagebox because the interaction with pressing a key is not
user friendly. You would have to click "ok" and then press a key which is not what we
want.

So from now we will be using info_Label to interact with the user.

on_set_key now packs this label then the user sets the key and we unpack it.